code analyzer fix

+++ BossMod/Modules/Stormblood/Trial/T05Yojimbo/DragonsLair.cs
@@ -36,7 +36,7 @@ class DragonsLair(BossModule module) : Components.Exaflare(module, _rect)
     }
     public override void DrawArenaForeground(int pcSlot, Actor pc)
     {
-        foreach (var head in Module.Enemies(OID.DragonsHead))
+        foreach (var head in Module.Enemies((uint)OID.DragonsHead))
             Arena.Actor(head, Colors.Object, true);
     }
 }
